requirements.txt
asgiref==3.11.0
Django==6.0.1
django-filter==25.2
djangorestframework==3.16.1
ruff==0.14.13
sqlparse==0.5.5
